common-apps:  update copy of *_nano.a

+++ common-apps-functions-source.sh
@@ -2545,14 +2545,29 @@ function copy_cross_nano_libs()
   local src_folder="$1"
   local dst_folder="$2"
 
+  # Copy the nano variants with a distinct name, as used in nano.specs.
+  cp -v -f "${src_folder}/libc.a" "${dst_folder}/libc_nano.a"
+  cp -v -f "${src_folder}/libg.a" "${dst_folder}/libg_nano.a"
+  cp -v -f "${src_folder}/libm.a" "${dst_folder}/libm_nano.a"
+
+
   cp -v -f "${src_folder}/libstdc++.a" "${dst_folder}/libstdc++_nano.a"
   cp -v -f "${src_folder}/libsupc++.a" "${dst_folder}/libsupc++_nano.a"
 
-  cp -v -f "${src_folder}/libc.a" "${dst_folder}/libc_nano.a"
-  cp -v -f "${src_folder}/libg.a" "${dst_folder}/libg_nano.a"
+  if [ -f "${src_folder}/libgloss.a" ]
+  then
+    cp -v -f "${src_folder}/libgloss.a" "${dst_folder}/libgloss_nano.a"
+  fi
 
-  cp -v -f "${src_folder}/librdimon.a" "${dst_folder}/librdimon_nano.a"
-  cp -v -f "${src_folder}/librdimon-v2m.a" "${dst_folder}/lrdimon-v2m_nano.a"
+  if [ -f "${src_folder}/librdimon.a" ]
+  then
+    cp -v -f "${src_folder}/librdimon.a" "${dst_folder}/librdimon_nano.a"
+  fi
+
+  if [ -f "${src_folder}/librdimon-v2m.a" ]
+  then
+    cp -v -f "${src_folder}/librdimon-v2m.a" "${dst_folder}/lrdimon-v2m_nano.a"
+  fi
 }
 
 # Copy target libraries from each multilib folders.
